requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
idna==2.10
Naked==0.1.31
pycryptodome==3.9.9
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
sqlparse==0.4.1
urllib3==1.25.11
